@@ -8,7 +8,7 @@ use PDO,
     Respect\Data\Styles\NorthWind,
     Respect\Relational\Mapper;
 
-class NorthWindTest extends \PHPUnit_Framework_TestCase
+class NorthWindTest extends \PHPUnit\Framework\TestCase
 {
 
     /**
@@ -26,7 +26,13 @@ class NorthWindTest extends \PHPUnit_Framework_TestCase
      */
     private $conn;
 
-    public function setUp()
+    private $posts;
+    private $authors;
+    private $comments;
+    private $categories;
+    private $postsCategories;
+
+    protected function setUp(): void
     {
 
         $conn = new PDO('sqlite::memory:');
@@ -150,7 +156,7 @@ class NorthWindTest extends \PHPUnit_Framework_TestCase
         $this->mapper->entityNamespace = __NAMESPACE__ . '\\';
     }
 
-    public function tableEntityProvider()
+    public static function tableEntityProvider()
     {
         return array(
             array('Posts',              'Posts'),
@@ -161,7 +167,7 @@ class NorthWindTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function manyToMantTableProvider()
+    public static function manyToMantTableProvider()
     {
         return array(
             array('Posts',  'Categories',   'PostCategories'),
@@ -170,7 +176,7 @@ class NorthWindTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function columnsPropertyProvider()
+    public static function columnsPropertyProvider()
     {
         return array(
             array('Text'),
@@ -181,7 +187,7 @@ class NorthWindTest extends \PHPUnit_Framework_TestCase
         );
     }
     
-    public function keyProvider()
+    public static function keyProvider()
     {
         return array(
             array('Posts',      'PostID'),
